EIP1-4179 Refactor Anonymous Elector Document data model to not have print requests - each new AED request creates a new AED document.
