+++ src/main/kotlin/uk/gov/dluhc/printapi/database/entity/AnonymousElectorDocumentStatus.kt
@@ -9,17 +9,14 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import uk.gov.dluhc.printapi.database.repository.UUIDCharType
 import uk.gov.dluhc.printapi.database.repository.UseExistingOrGenerateUUID
 import java.time.Instant
-import java.time.LocalDate
 import java.util.UUID
-import javax.persistence.CascadeType
 import javax.persistence.Entity
 import javax.persistence.EntityListeners
-import javax.persistence.FetchType
+import javax.persistence.EnumType
+import javax.persistence.Enumerated
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
 import javax.persistence.Id
-import javax.persistence.JoinColumn
-import javax.persistence.OneToMany
 import javax.persistence.Table
 import javax.persistence.Version
 import javax.validation.constraints.NotNull
@@ -28,7 +25,7 @@ import javax.validation.constraints.Size
 @Table
 @Entity
 @EntityListeners(AuditingEntityListener::class)
-class AedPrintRequest(
+class AnonymousElectorDocumentStatus(
     @Id
     @Type(type = UUIDCharType)
     @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "UUID")
@@ -36,26 +33,11 @@ class AedPrintRequest(
     var id: UUID? = null,
 
     @field:NotNull
-    @field:Size(max = 7)
-    var electoralRollNumber: String,
+    @Enumerated(EnumType.STRING)
+    var status: Status,
 
     @field:NotNull
-    @field:Size(max = 255)
-    var aedTemplateFilename: String,
-
-    @field:NotNull
-    var issueDate: LocalDate,
-
-    @field:NotNull
-    var requestDateTime: Instant,
-
-    @field:NotNull
-    @field:Size(max = 255)
-    var userId: String,
-
-    @OneToMany(cascade = [CascadeType.ALL], orphanRemoval = true, fetch = FetchType.EAGER)
-    @JoinColumn(name = "aed_print_request_id", nullable = false)
-    var statusHistory: MutableList<AedPrintRequestStatus> = mutableListOf(),
+    var eventDateTime: Instant,
 
     @CreationTimestamp
     var dateCreated: Instant? = null,
@@ -66,21 +48,11 @@ class AedPrintRequest(
 
     @Version
     var version: Long? = null
-
 ) {
-
-    fun addPrintRequestStatus(newPrintRequestStatus: AedPrintRequestStatus): AedPrintRequest {
-        statusHistory += newPrintRequestStatus
-        return this
-    }
-
-    val status: AedPrintRequestStatus.Status?
-        get() = statusHistory.sortedByDescending { it.eventDateTime }.first().status
-
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other == null || Hibernate.getClass(this) != Hibernate.getClass(other)) return false
-        other as AedPrintRequest
+        other as AnonymousElectorDocumentStatus
 
         return id != null && id == other.id
     }
@@ -89,6 +61,11 @@ class AedPrintRequest(
 
     @Override
     override fun toString(): String {
-        return this::class.simpleName + "(id = $id , dateCreated = $dateCreated , createdBy = $createdBy , version = $version )"
+        return this::class.simpleName + "(id = $id , status = $status , dateCreated = $dateCreated , createdBy = $createdBy , version = $version )"
+    }
+
+    enum class Status {
+        GENERATED,
+        PRINTED,
     }
 }
