+++ src/main/kotlin/uk/gov/dluhc/printapi/service/aed/AedPdfTemplateDetailsFactory.kt
@@ -3,7 +3,6 @@ package uk.gov.dluhc.printapi.service.aed
 import org.springframework.stereotype.Component
 import software.amazon.awssdk.services.s3.S3Client
 import uk.gov.dluhc.printapi.config.AnonymousElectorDocumentPdfTemplateProperties
-import uk.gov.dluhc.printapi.database.entity.AedPrintRequest
 import uk.gov.dluhc.printapi.database.entity.AnonymousElectorDocument
 import uk.gov.dluhc.printapi.service.isWalesCode
 import uk.gov.dluhc.printapi.service.pdf.AbstractElectorDocumentPdfTemplateDetailsFactory
@@ -19,33 +18,32 @@ class AedPdfTemplateDetailsFactory(
         buildDetails(electorDocument.gssCode, electorDocument.photoLocationArn, getPlaceholders(electorDocument))
 
     private fun getPlaceholders(electorDocument: AnonymousElectorDocument): Map<String, String> {
-        val printRequest = electorDocument.getLatestPrintRequest()
         return if (isWalesCode(electorDocument.gssCode)) {
-            getWelshTemplatePlaceholders(electorDocument, printRequest)
+            getWelshTemplatePlaceholders(electorDocument)
         } else {
-            getEnglishTemplatePlaceholders(electorDocument, printRequest)
+            getEnglishTemplatePlaceholders(electorDocument)
         }
     }
 
-    private fun getWelshTemplatePlaceholders(
-        electorDocument: AnonymousElectorDocument,
-        printRequest: AedPrintRequest
-    ): Map<String, String> {
-        return mapOf(
-            pdfTemplateProperties.welsh.placeholder.electoralRollNumber to printRequest.electoralRollNumber.uppercase(),
-            pdfTemplateProperties.welsh.placeholder.dateOfIssue to printRequest.issueDate.format(DATE_TIME_FORMATTER),
-            pdfTemplateProperties.welsh.placeholder.certificateNumber to electorDocument.certificateNumber,
-        )
+    private fun getWelshTemplatePlaceholders(electorDocument: AnonymousElectorDocument): Map<String, String> {
+        return with(electorDocument) {
+            mapOf(
+                pdfTemplateProperties.welsh.placeholder.electoralRollNumber to electoralRollNumber.uppercase(),
+                pdfTemplateProperties.welsh.placeholder.dateOfIssue to issueDate.format(DATE_TIME_FORMATTER),
+                pdfTemplateProperties.welsh.placeholder.certificateNumber to certificateNumber,
+            )
+        }
     }
 
     private fun getEnglishTemplatePlaceholders(
-        electorDocument: AnonymousElectorDocument,
-        printRequest: AedPrintRequest
+        electorDocument: AnonymousElectorDocument
     ): Map<String, String> {
-        return mapOf(
-            pdfTemplateProperties.english.placeholder.electoralRollNumber to printRequest.electoralRollNumber.uppercase(),
-            pdfTemplateProperties.english.placeholder.dateOfIssue to printRequest.issueDate.format(DATE_TIME_FORMATTER),
-            pdfTemplateProperties.english.placeholder.certificateNumber to electorDocument.certificateNumber,
-        )
+        return with(electorDocument) {
+            mapOf(
+                pdfTemplateProperties.english.placeholder.electoralRollNumber to electoralRollNumber.uppercase(),
+                pdfTemplateProperties.english.placeholder.dateOfIssue to issueDate.format(DATE_TIME_FORMATTER),
+                pdfTemplateProperties.english.placeholder.certificateNumber to certificateNumber,
+            )
+        }
     }
 }
