+++ src/main/resources/db/changelog/db.changelog-master.xml
@@ -21,4 +21,5 @@
     <include relativeToChangelogFile="true" file="ddl/0012_supporting_information_non_mandatory.xml"/>
     <include relativeToChangelogFile="true" file="ddl/0013_initial_retention_data_removed_column.xml"/>
     <include relativeToChangelogFile="true" file="ddl/0014_create_anonymous_elector_document_tables.xml"/>
+    <include relativeToChangelogFile="true" file="ddl/0015_refactor_anonymous_elector_document_tables.xml"/>
 </databaseChangeLog>
