+++ src/test/kotlin/uk/gov/dluhc/printapi/testsupport/testdata/TestFixedDataFactory.kt
@@ -1,7 +1,7 @@
 package uk.gov.dluhc.printapi.testsupport.testdata
 
 import uk.gov.dluhc.printapi.database.entity.AddressFormat
-import uk.gov.dluhc.printapi.database.entity.AedPrintRequestStatus
+import uk.gov.dluhc.printapi.database.entity.AnonymousElectorDocumentStatus
 import uk.gov.dluhc.printapi.database.entity.DeliveryAddressType
 import uk.gov.dluhc.printapi.database.entity.DeliveryClass
 import uk.gov.dluhc.printapi.database.entity.PrintRequestStatus.Status
@@ -34,7 +34,7 @@ fun aValidAnonymousElectorDocumentTemplateFilename(): String = "anonymous-electo
 
 fun aValidTemporaryCertificateStatus() = TemporaryCertificateStatus.Status.GENERATED
 
-fun aValidAnonymousElectorDocumentStatus() = AedPrintRequestStatus.Status.GENERATED
+fun aValidAnonymousElectorDocumentStatus() = AnonymousElectorDocumentStatus.Status.GENERATED
 
 fun aValidSourceTypeDto() = SourceTypeDto.VOTER_CARD
